backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-dotenv
groq
httpx
pydantic
jinja2
python-docx